lint/unconvert: remove unnecessary conversions
The main one is that system.Copy (unlike io.Copy) returns int64, so we
don't need to convert to int64. The 64-bit conversions in updateHeader
are necessary because on 32-bit systems you will get build failures
without the conversions.

+++ oci/cas/dir/dir.go
@@ -191,7 +191,7 @@ func (e *dirEngine) PutBlob(ctx context.Context, reader io.Reader) (_ digest.Dig
 		return "", -1, fmt.Errorf("rename temporary blob: %w", err)
 	}
 
-	return digester.Digest(), int64(size), nil
+	return digester.Digest(), size, nil
 }
 
 // GetBlob returns a reader for retrieving a blob from the image, which the
